+++ src/Telegram.Bot/TelegramBotClient.cs
@@ -288,14 +288,66 @@ namespace Telegram.Bot
 
             var fileUri = new Uri($"{_baseFileUrl}{filePath}", UriKind.Absolute);
 
-            using var response = await _httpClient
-                .GetAsync(fileUri, HttpCompletionOption.ResponseHeadersRead, cancellationToken)
-                .ConfigureAwait(false);
+            HttpResponseMessage? httpResponse;
+            try
+            {
+                httpResponse = await _httpClient
+                    .GetAsync(fileUri, HttpCompletionOption.ResponseHeadersRead, cancellationToken)
+                    .ConfigureAwait(false);
+            }
+            catch (TaskCanceledException exception)
+            {
+                if (cancellationToken.IsCancellationRequested)
+                {
+                    throw;
+                }
 
-            response.EnsureSuccessStatusCode();
+                throw new RequestException("Request timed out", exception);
+            }
+            catch (Exception exception)
+            {
+                throw new RequestException("Exception during file download", exception);
+            }
 
-            await response.Content.CopyToAsync(destination)
-                .ConfigureAwait(false);
+            try
+            {
+                if (!httpResponse.IsSuccessStatusCode)
+                {
+                    var failedApiResponse = await httpResponse
+                        .DeserializeContentAsync<FailedApiResponse>(includeBody: false)
+                        .ConfigureAwait(false);
+
+                    throw ExceptionParser.Parse(
+                        failedApiResponse.ErrorCode,
+                        failedApiResponse.Description,
+                        failedApiResponse.Parameters
+                    );
+                }
+
+                if (httpResponse.Content is null)
+                    throw new RequestException(
+                        "Response doesn't contain any content",
+                        httpResponse.StatusCode
+                    );
+
+                try
+                {
+                    await httpResponse.Content.CopyToAsync(destination)
+                        .ConfigureAwait(false);
+                }
+                catch (Exception exception)
+                {
+                    throw new RequestException(
+                        "Exception during file download",
+                        httpResponse.StatusCode,
+                        exception
+                    );
+                }
+            }
+            finally
+            {
+                httpResponse?.Dispose();
+            }
         }
 
         /// <inheritdoc />