+++ CHANGELOG.md
@@ -65,6 +65,9 @@ except `TaskCancelledException` that is triggered by token cancellation and erro
 `MakingApiRequest` and `ApiResponseReceived` event handlers
 - Interface `IChatMessage` renamed into `IChatTargetable`
 - Requests with a property `long ChatId { get; }` explicitly implement `IChatTargetable`
+- Changed error handling in `DownloadFileAsync`: `ApiRequestException` will be thrown when 
+correct error response is received from Bot API, otherwise `RequestException` is thrown,
+`TaskCancelledException` is thrown when cancellation is requested 
 
 ### Removed
 - Enum member `ParseMode.Default`